renamed the updater package to fix
