+++ buffalo/cmd/fix.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/gobuffalo/buffalo/buffalo/cmd/updater"
+	"github.com/gobuffalo/buffalo/buffalo/cmd/fix"
 	"github.com/gobuffalo/buffalo/runtime"
 	"github.com/spf13/cobra"
 )
@@ -14,7 +14,7 @@ var fixCmd = &cobra.Command{
 	Aliases: []string{"update"},
 	Short:   fmt.Sprintf("will attempt to fix a Buffalo application's API to match version %s", runtime.Version),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		return updater.Run()
+		return fix.Run()
 	},
 }
