+++ buffalo/cmd/fix/imports.go
@@ -1,4 +1,4 @@
-package updater
+package fix
 
 import (
 	"bytes"
