+++ buffalo/cmd/fix/npm.go
@@ -1,4 +1,4 @@
-package updater
+package fix
 
 import (
 	"bytes"
